Refactored Unit testing structure for consistency Refactored unit testing structure and test locations to be consistent with style of having unit testing folder structure reproduce source code module structure, with each testing file testing one method. Update README to detail test naming styling consistent with this style.
