+++ LDAR_Sim/testing/README.md
@@ -1,6 +1,32 @@
 
 # Unit Testing Info
 
+## Unit testing file format guide
+
+Inside of the unit_testing folder, tests should be grouped into sub-folders based on the module that is being tested, with folders matching LDAR-SIm module hierarchy. The folder should be named test_\<module name> For example: testing the method detect_emissions in crews.py should results a file named test_detect_emissions inside a folder structure of test_methods -> test_crews -> test_detect_emissions.py. Each python file containing tests should correspond to a specific method.
+
+## Test naming guide
+
+Tests should be named descriptively, as the test names will not need to be called. Tests must start with test_\<xxx>_ where xxx is a unique number corresponding to all tests in the given testing folder, where the first two numbers are used to indicate the higher level module path, and the final number is used to indicate each test file corresponding to a method within the module. For example: a test testing the method detect_emissions in crews.py should start with test_05x where x is the number corresponding to methods in crews.py
+
+Module Unique Identifiers have been outlined below:
+
+- Files directly under src: 00
+- campaigns: 01
+- economics: 02
+- geography: 03
+- initialization: 04
+- methods: 05
+  - deployment: 11
+  - init_func: 12
+  - reporting: 13
+  - sensors: 14
+- out_processing: 06
+- utils: 07
+- weather: 08
+
+## Required installation
+
 Before running Pytest coverage, install
 
 ```cmd
