+++ LDAR_Sim/testing/unit_testing/test_methods/test_crew/test_detect_emissions.py
@@ -7,8 +7,7 @@ from src.methods.crew import BaseCrew
 import numpy as np
 from pathlib import Path
 import sys
-
-from testing.unit_testing.test_methods.crew_testing_fixutres import (                    # noqa: F401
+from testing.unit_testing.test_methods.test_crew.crew_testing_fixtures import (     # noqa: F401
                                     mock_site_for_crew_testing_1_fix,
                                     mock_config_for_crew_testing_1_fix,
                                     mock_params_for_crew_testing_1_fix,
@@ -18,13 +17,14 @@ from testing.unit_testing.test_methods.crew_testing_fixutres import (           
                                     mock_state_for_crew_testing_2_fix)
 
 
-def test_004_detect_emissions_simple(mock_config_for_crew_testing_1,
+def test_021_detect_emissions_simple(mock_config_for_crew_testing_1,
                                      mock_site_for_crew_testing_1,
                                      mock_params_for_crew_testing_1,
                                      mocker) -> None:
     np.random.seed(0)  # Setting a seed for reproducibility
+    # Add src directory to the path
     sys.path.insert(1, str(Path(os.path.dirname(os.path.realpath(__file__))
-                                ).parent.parent.parent / "src"))  # Add src directory to the path
+                                ).parent.parent.parent.parent / "src"))
     # Mock the sensors detect emissions function called inside of crew.detect_emissions
     # to return the arguments it's been passed. This way we can check the behavior of
     # crew.detect emissions without being dependant on sensor.detect_emissions.
@@ -41,14 +41,15 @@ def test_004_detect_emissions_simple(mock_config_for_crew_testing_1,
     assert result == expected
 
 
-def test_004_detect_emissions_simple_w_vents(mock_state_for_crew_testing_2,
+def test_021_detect_emissions_simple_w_vents(mock_state_for_crew_testing_2,
                                              mock_config_for_crew_testing_2,
                                              mock_site_for_crew_testing_2,
                                              mock_params_for_crew_testing_2,
                                              mocker) -> None:
     np.random.seed(0)  # Setting a seed for reproducibility
+    # Add src directory to the path
     sys.path.insert(1, str(Path(os.path.dirname(os.path.realpath(__file__))
-                                ).parent.parent.parent / "src"))  # Add src directory to the path
+                                ).parent.parent.parent.parent / "src"))
     # Mock the sensors detect emissions function called inside of crew.detect_emissions
     # to return the arguments it's been passed. This way we can check the behavior of
     # crew.detect emissions without being dependant on sensor.detect_emissions.
