+++ LDAR_Sim/testing/unit_testing/test_methods/test_funcs/test_measured_rate.py
@@ -1,4 +1,4 @@
-"""Unit test for testing measured rate"""
+"""Unit test for testing functions.py"""
 
 import pytest
 import numpy as np
@@ -10,7 +10,7 @@ from src.methods.funcs import measured_rate
     ([10, 0.5], 18.8),
     ([8, 0.3], 12.2),
 ])
-def test_003_measured_rate(test_input, expected):
+def test_001_measured_rate(test_input, expected):
     np.random.seed(0)  # Setting a seed for reproducibility
     result = measured_rate(test_input[0], test_input[1])
     assert np.isclose(
