+++ LDAR_Sim/testing/unit_testing/test_utils/test_generic_functions/test_geo_idx.py
@@ -3,7 +3,7 @@ import numpy as np
 from src.utils.generic_functions import geo_idx
 
 
-def test_001_geo_idx_valid():
+def test_051_geo_idx_valid():
     """ Test geo_idx - Valid """
     dd = 10.123
     dd_array = np.array([9.0, 10.0, 11.0, 12.0, 13.0])
